+++ src/main/java/com/homies/app/service/UserService.java
@@ -14,8 +14,8 @@ import java.time.temporal.ChronoUnit;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
-
 import com.homies.app.service.dto.UserUpdateDTO;
+import com.homies.app.service.errors.EmailAlreadyUsedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.cache.CacheManager;
